feat(app entry point): configure app graphql entry point
